Remove invalid item from depends field of library.properties
The depends field of the library.properties metadata file specifies the dependencies that should be installed along with
the library by the Arduino Library Manager.

This field must contain only the names of libraries that are available for installation via Library Manager.

The presence of any items which are not in Library Manager causes installation of the library to fail:

- Arduino IDE 1.x: "no protocol:" error
- Arduino IDE 2.x: fails silently
- Arduino CLI: "No valid dependencies solution found" error

My guess is that this was intended to be "Bolder Flight Systems Unit Conversions". However, I was not able to verify this because I found the "Bolder Flight Systems UBLOX" library and its example sketch don't have any dependencies on other Arduino libraries. For this reason, I opted to remove the invalid item from the `depends` field entirely instead of correcting the name.